Change price to bigdecimal and redo price calculation

+++ Dine-A-Bite_BE/OrderManager/src/main/java/cobol/services/ordermanager/domain/entity/Food.java
@@ -30,7 +30,7 @@ public class Food {
     private int stock;
 
     private String description;
-    private float price;
+    private BigDecimal price;
     private int preparationTime;
 
     @ManyToMany(fetch = FetchType.EAGER)
@@ -80,7 +80,7 @@ public class Food {
 
         //Setting general fields
         this.description = cf.getDescription();
-        this.price = cf.getPrice().floatValue();
+        this.price = cf.getPrice();
         this.preparationTime = cf.getPreparationTime();
         this.stock = cf.getStock();
 
@@ -100,7 +100,7 @@ public class Food {
 
         //Setting general fields
         this.description = cf.getDescription().equals("") ? this.description : cf.getDescription();
-        this.price = cf.getPrice().floatValue() < 0 ? this.price : cf.getPrice().floatValue();
+        this.price = cf.getPrice().floatValue() < 0 ? this.price : cf.getPrice();
         this.preparationTime = cf.getPreparationTime() < 0 ? this.preparationTime : cf.getPreparationTime();
         //Stock will be 0 if no stock has changed
         this.stock += cf.getStock();
@@ -125,7 +125,7 @@ public class Food {
      */
     public Food(CommonFood cf, Stand stand) {
         this.description = cf.getDescription();
-        this.price = cf.getPrice().floatValue();
+        this.price = cf.getPrice();
         this.preparationTime = cf.getPreparationTime();
         this.stock = cf.getStock();
         this.foodId = new FoodId(cf.getName(), stand);
@@ -145,7 +145,7 @@ public class Food {
     public CommonFood asCommonFood() {
         return new CommonFood(
                 foodId.name,
-                BigDecimal.valueOf(price),
+                price,
                 preparationTime,
                 stock,
                 foodId.stand.getName(),
@@ -288,11 +288,11 @@ public class Food {
         this.description = description;
     }
 
-    public float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
